project/requirements.txt
asgiref==3.8.1
Django==5.1.3
paho-mqtt==2.1.0
sqlparse==0.5.2
paho-mqtt==2.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
